feat: Mas endpoints, falta hacer filtros
