+++ models/server.js
@@ -18,6 +18,7 @@ class Server {
         }); // Integrante 2 */
     this.app.use('/api/v1/clima', require('../routes/clima'))
     this.app.use('/api/v1/aire', require('../routes/aire'))
+    this.app.use('/api/v1/pronostico', require('../routes/pronostico'))
   }
 
   listen () {
