+++ routes/aire.js
@@ -1,7 +1,9 @@
 const { Router } = require('express')
 const rutas = Router()
-const { getPolucionAire } = require('../controllers/aire')
+const { getPolucionAire, getPolucionAireHistorica, getPolucionAireExtendida } = require('../controllers/aire')
 
 rutas.get('/polucion', getPolucionAire)
+rutas.get('/polucionHistorica', getPolucionAireHistorica)
+rutas.get('/polucionExtendida', getPolucionAireExtendida)
 
 module.exports = rutas
